requirements.txt
appdirs==1.4.3
asgiref==3.2.7
astroid==2.3.3
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
distlib==0.3.0
Django==3.0.5
django-cors-headers==3.2.1
django-extensions==2.2.9
django-rest-framework==0.1.0
djangorestframework==3.11.0
filelock==3.0.12
gunicorn==20.0.4
idna==2.9
isort==4.3.21
jedi==0.17.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
parso==0.7.0
prompt-toolkit==3.0.5
ptpython==3.0.2
Pygments==2.6.1
pylint==2.4.4
pytz==2019.3
requests==2.23.0
six==1.14.0
sqlparse==0.3.1
stripe==2.44.0
urllib3==1.25.8
virtualenv==20.0.7
wcwidth==0.1.9
whitenoise==5.0.1
wrapt==1.11.2
